Move Tenable API types
