+++ src/converters/AccountContainerRelationshipConverter.test.ts
@@ -1,4 +1,5 @@
-import { Account, Container } from "../types";
+import { Container } from "../tenable/types";
+import { Account } from "../types";
 import { createAccountContainerRelationships } from "./AccountContainerRelationshipConverter";
 
 test("convert account container relationship", () => {
