+++ src/converters/AssetWebAppVulnerabilityRelationshipConverter.test.ts
@@ -1,4 +1,4 @@
-import { Dictionary } from "../types";
+import { Dictionary } from "../tenable/types";
 import { createAssetWebAppVulnerabilityRelationships } from "./AssetWebAppVulnerabilityRelationshipConverter";
 
 test("convert application vulnerability relationships", () => {
