+++ src/converters/AssetWebAppVulnerabilityRelationshipConverter.ts
@@ -6,7 +6,7 @@ import {
   WEBAPP_VULNERABILITY_ENTITY_TYPE,
 } from "../jupiterone/entities";
 
-import { Asset, Dictionary, WebAppVulnerability } from "../types";
+import { Asset, Dictionary, WebAppVulnerability } from "../tenable/types";
 import {
   generateEntityKey,
   generateRelationshipKey,
