+++ src/converters/MalwareEntityConverter.ts
@@ -3,11 +3,11 @@ import {
   MALWARE_ENTITY_TYPE,
   MalwareVulnerabilityEntity,
 } from "../jupiterone/entities";
-import { Dictionary, Malware } from "../types";
+import { ContainerMalware, Dictionary } from "../tenable/types";
 import { generateEntityKey } from "../utils/generateKey";
 
 export function createMalwareEntities(
-  data: Dictionary<Malware[]>,
+  data: Dictionary<ContainerMalware[]>,
 ): MalwareVulnerabilityEntity[] {
   const defaultValue: MalwareVulnerabilityEntity[] = [];
   const vulnerabilityArrays = Object.values(data);
@@ -26,7 +26,7 @@ export function createMalwareEntities(
 }
 
 function createMalwareEntity(
-  vulnerability: Malware,
+  vulnerability: ContainerMalware,
 ): MalwareVulnerabilityEntity {
   const malwareId = vulnerability.md5;
   return {
