+++ src/converters/ReportMalwareRelationshipConverter.test.ts
@@ -1,4 +1,4 @@
-import { Dictionary, Malware, Report } from "../types";
+import { ContainerMalware, Dictionary, Report } from "../tenable/types";
 import { createReportMalwareRelationships } from "./ReportMalwareRelationshipConverter";
 
 test("convert report container vulnerability relationship", () => {
@@ -23,7 +23,7 @@ test("convert report container vulnerability relationship", () => {
       os_architecture: "AMD64",
     },
   ];
-  const malwares: Dictionary<Malware[]> = {
+  const malwares: Dictionary<ContainerMalware[]> = {
     "sha256:c42a932fda50763cb2a0169dd853f071a37629cfa4a477b81b4ee87c2b0bb3dc": [
       {
         infectedFile: "string",
