+++ src/converters/ReportMalwareRelationshipConverter.ts
@@ -5,7 +5,7 @@ import {
   REPORT_MALWARE_RELATIONSHIP_TYPE,
   ReportMalwareRelationship,
 } from "../jupiterone/entities";
-import { Dictionary, Malware, Report } from "../types";
+import { ContainerMalware, Dictionary, Report } from "../tenable/types";
 import {
   generateEntityKey,
   generateRelationshipKey,
@@ -13,7 +13,7 @@ import {
 
 export function createReportMalwareRelationships(
   reports: Report[],
-  malwares: Dictionary<Malware[]>,
+  malwares: Dictionary<ContainerMalware[]>,
 ): ReportMalwareRelationship[] {
   const defaultValue: ReportMalwareRelationship[] = [];
   const relationships = reports.reduce((acc, report) => {
@@ -27,7 +27,7 @@ export function createReportMalwareRelationships(
 }
 
 function createRelation(
-  vulnerability: Malware,
+  vulnerability: ContainerMalware,
   reportId: string,
 ): ReportMalwareRelationship {
   const malwareId = vulnerability.md5;
