+++ src/converters/ScanWebAppVulnerabilityRelationshipConverter.test.ts
@@ -1,4 +1,4 @@
-import { Dictionary } from "../types";
+import { Dictionary } from "../tenable/types";
 import { createScanWebAppVulnerabilityRelationships } from "./ScanWebAppVulnerabilityRelationshipConverter";
 
 const vulnerabilitiesDictionary: Dictionary<any[]> = {
@@ -116,13 +116,9 @@ const scans: any[] = [
         targets: "https://dualbootpartners.com",
         alt_targets_used: false,
         "pci-can-upload": false,
-        scan_start: 1555066123,
         timestamp: 1555068471,
-        scan_end: 1555068471,
         haskb: false,
         hasaudittrail: true,
-        scanner_start: null,
-        scanner_end: null,
       },
       hosts: [
         {
