+++ src/converters/ScanWebAppVulnerabilityRelationshipConverter.ts
@@ -5,7 +5,7 @@ import {
   ScanWebAppVulnerabilityRelationship,
   WEBAPP_VULNERABILITY_ENTITY_TYPE,
 } from "../jupiterone/entities";
-import { Dictionary, Scan, WebAppVulnerability } from "../types";
+import { Dictionary, Scan, WebAppVulnerability } from "../tenable/types";
 import {
   generateEntityKey,
   generateRelationshipKey,
