+++ src/converters/UserEntityConverter.test.ts
@@ -1,4 +1,4 @@
-import { User } from "../types";
+import { User } from "../tenable/types";
 import { createUserEntities } from "./UserEntityConverter";
 
 test("convert user entity", () => {
