+++ src/converters/WebAppVulnerabilityEntityConverter.test.ts
@@ -1,4 +1,4 @@
-import { Dictionary } from "../types";
+import { Dictionary } from "../tenable/types";
 import { createVulnerabilityEntities } from "./WebAppVulnerabilityEntityConverter";
 
 test("convert vulnerability entities", () => {
