+++ src/converters/WebAppVulnerabilityEntityConverter.ts
@@ -3,7 +3,7 @@ import {
   WEBAPP_VULNERABILITY_ENTITY_TYPE,
   WebAppVulnerabilityEntity,
 } from "../jupiterone/entities";
-import { Dictionary, WebAppVulnerability } from "../types";
+import { Dictionary, WebAppVulnerability } from "../tenable/types";
 import { generateEntityKey } from "../utils/generateKey";
 
 export function createVulnerabilityEntities(
