+++ src/tenable/TenableClient.test.ts
@@ -1,6 +1,7 @@
 import nock from "nock";
 import { fetchTenableData } from "./index";
 import TenableClient from "./TenableClient";
+import { Scan } from "./types";
 
 const ACCESS_KEY =
   process.env.TENABLE_LOCAL_EXECUTION_ACCESS_KEY || "test_access_token";
@@ -60,7 +61,7 @@ describe("TenableClient fetch ok data", () => {
     const { nockDone } = await nock.back("scan-ok.json", {
       before: prepareScope,
     });
-    const scan = await client.fetchScanById(12);
+    const scan = await client.fetchScanDetail({ id: 12 } as Scan);
     nockDone();
 
     expect(scan).not.toEqual({});
